chore: 更新脚本和文档提示

+++ configure.py
@@ -6,6 +6,11 @@ assets_dir = Path(__file__).parent / "assets"
 
 
 def configure_ocr_model():
+    if not (assets_dir / "MaaCommonAssets" / "OCR").exists():
+        print("Please clone this repository completely, don’t miss \"--recursive\", and don’t download the zip package!")
+        print("请完整克隆本仓库，不要漏掉 \"--recursive\"，也不要下载 zip 包！")
+        exit(1)
+
     shutil.copytree(
         assets_dir / "MaaCommonAssets" / "OCR" / "ppocr_v4" / "zh_cn",
         assets_dir / "resource" / "base" / "model" / "ocr",
@@ -15,3 +20,5 @@ def configure_ocr_model():
 
 if __name__ == "__main__":
     configure_ocr_model()
+
+    print("OCR model configured.")
