+++ install.py
@@ -13,6 +13,11 @@ version = len(sys.argv) > 1 and sys.argv[1] or "v0.0.1"
 
 
 def install_deps():
+    if not (working_dir / "deps" / "bin").exists():
+        print("Please download the MaaFramework to \"deps\" first.")
+        print("请先下载 MaaFramework 到 \"deps\"。")
+        sys.exit(1)
+
     shutil.copytree(
         working_dir / "deps" / "bin",
         install_path,
@@ -69,3 +74,5 @@ if __name__ == "__main__":
     install_deps()
     install_resource()
     install_chores()
+
+    print(f"Install to {install_path} successfully.")
